[bsp][hpmicro] add HPM5300EVK,HPM5301EVKLITE and HPM6800EVK support & update hpm_sdk
- added new boards: hpm5300evk, hpm5301evklite and hpm6800evk
- upgaded hpm_sdk
- driver updates and bugfixes
- add hpmicro BSPs to CI

+++ bsp/hpmicro/libraries/hpm_sdk/drivers/src/hpm_rng_drv.c
@@ -69,7 +69,7 @@ static hpm_stat_t rng_rand(RNG_Type *ptr, void *buf, uint32_t count_in_byte, boo
         } while (!fifo_level && wait);
 
         if (fifo_level) {
-            *(uint32_t *)(buf + i * sizeof(uint32_t)) = ptr->FO2B;
+            *(uint32_t *)((uint32_t)buf + i * sizeof(uint32_t)) = ptr->FO2B;
         } else {
             stat = status_rng_not_available;
             break;
